requirements.txt
streamlit
gradio
openai
anthropic
pandas
numpy
matplotlib
requests
newspaper3k
python-dotenv
scikit-learn
lxml_html_clean
